make setting field in own module
